educa/courses/api/permissions.py
from rest_framework.permissions import BasePermission


class IsEnrolled(BasePermission):
    def has_object_permission(self, request, view, obj):
        return obj.students.filter(id=request.user.id).exists()  # проверк чт в списке студентов курса есть request.user
