requirement.txt
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
certifi==2023.11.17
charset-normalizer==2.1.1
Deprecated==1.2.14
Django==4.1.13
django-braces==1.15.0
django-debug-toolbar==3.6.0
django-embed-video==1.4.4
django-redisboard==8.3.0
djangorestframework==3.13.1
idna==3.6
packaging==23.2
Pillow==9.2.0
pymemcache==3.5.2
pytz==2023.3.post1
redis==4.3.4
requests==2.28.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
urllib3==1.26.18
wrapt==1.16.0
